・バリデーションの追加

+++ backend/index.js
@@ -2,8 +2,6 @@ const express = require("express");
 const app = express();
 const port = process.env.PORT || 3001;
 const mysql = require("mysql2/promise");
-const { validationResult } = require("express-validator");
-const ItemRegistValidator = require("./itemRegistValidator");
 const itemRegistValidator = require("./itemRegistValidator");
 
 app.use(express.json());
@@ -12,15 +10,13 @@ app.use(express.urlencoded({ extended: true }));
 // アカウント作成
 app.post(
   "/signUpInfo",
-  ItemRegistValidator.signUp,
+  itemRegistValidator.signUp,
 
   async (req, res) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      // バリデーション失敗
-      return res.status(422).json({ errors: errors.array() });
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
     }
+
     try {
       const sql =
         "INSERT INTO USER(mail_address, user_name, password) VALUES(?, ?, ?)";
@@ -36,15 +32,13 @@ app.post(
 // ユーザ情報を取得
 app.get(
   "/getUserInfo",
-  ItemRegistValidator.getUser,
+  itemRegistValidator.getUser,
 
   async (req, res) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      // バリデーション失敗
-      return res.status(422).json({ errors: errors.array() });
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
     }
+
     try {
       const sql =
         "SELECT id,user_name FROM USER WHERE mail_address = ? AND password = ?";
@@ -101,12 +95,10 @@ app.post(
   itemRegistValidator.addTodo,
 
   async (req, res) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      // バリデーション失敗
-      return res.status(422).json({ errors: errors.array() });
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
     }
+
     try {
       const sql =
         "INSERT INTO TASK(title, content, priority, user_id, deadline) VALUES(?, ?, ?, ?, ?)";
@@ -126,40 +118,57 @@ app.post(
 );
 
 // Todoを削除
-app.delete("/deleteTodo/:todoId", async (req, res) => {
-  try {
-    const sql = "DELETE FROM TASK WHERE id = ?";
-    const placeholder = req.params.todoId;
-    const results = await executeQuery(sql, placeholder);
-    res.send(results);
-  } catch (err) {
-    res.status(500).send(err);
+app.delete(
+  "/deleteTodo/:todoId",
+  itemRegistValidator.deleteTodo,
+
+  async (req, res) => {
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
+    }
+
+    try {
+      const sql = "DELETE FROM TASK WHERE id = ?";
+      const placeholder = req.params.todoId;
+      const results = await executeQuery(sql, placeholder);
+      res.send(results);
+    } catch (err) {
+      res.status(500).send(err);
+    }
   }
-});
+);
 
 // Todoを全削除
-app.delete("/clearTodos/:userId", async (req, res) => {
-  try {
-    const sql = "DELETE FROM TASK WHERE user_id = ?";
-    const placeholder = req.params.userId;
-    const results = await executeQuery(sql, placeholder);
-    res.send(results);
-  } catch (err) {
-    res.status(500).send(err);
+app.delete(
+  "/clearTodos/:userId",
+  itemRegistValidator.clearTodos,
+
+  async (req, res) => {
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
+    }
+
+    try {
+      const sql = "DELETE FROM TASK WHERE user_id = ?";
+      const placeholder = req.params.userId;
+      const results = await executeQuery(sql, placeholder);
+      res.send(results);
+    } catch (err) {
+      res.status(500).send(err);
+    }
   }
-});
+);
 
 // 完了・未完了のステータス切り替え
 app.patch(
   "/completionChange",
-  itemRegistValidator.completisionChange,
-  async (req, res) => {
-    const errors = validationResult(req);
+  itemRegistValidator.completionChange,
 
-    if (!errors.isEmpty()) {
-      // バリデーション失敗
-      return res.status(422).json({ errors: errors.array() });
+  async (req, res) => {
+    if (itemRegistValidator.checkVaridationResult(req, res)) {
+      return;
     }
+
     try {
       const sql = "UPDATE TASK SET completion = ? WHERE id = ?";
       const placeholder = [req.body.completion, req.body.id];
@@ -173,12 +182,10 @@ app.patch(
 
 // Todoを更新
 app.put("/editTodo/:id", itemRegistValidator.editTodo, async (req, res) => {
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    // バリデーション失敗
-    return res.status(422).json({ errors: errors.array() });
+  if (itemRegistValidator.checkVaridationResult(req, res)) {
+    return;
   }
+
   try {
     const sql =
       "UPDATE TASK SET title=?, content=?, priority=?, user_id=?, deadline=? WHERE id=?";
